% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeBan.R
\name{barCode_ban_add}
\alias{barCode_ban_add}
\title{读取数据添加到隔离区}
\usage{
barCode_ban_add(
  file = "data-raw/条码隔离区下载模板.xlsx",
  conn = tsda::conn_rds("LCERP")
)
}
\arguments{
\item{file}{文件}

\item{conn}{连接}
}
\value{
返回值
}
\description{
读取数据添加到隔离区
}
\examples{
barCode_ban_add()
}
