% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeBan.R
\name{barcode_ban_readFile}
\alias{barcode_ban_readFile}
\title{读取条码隔离模板信息}
\usage{
barcode_ban_readFile(file = "data-raw/条码隔离区下载模板.xlsx")
}
\arguments{
\item{file}{文件}
}
\value{
返回值
}
\description{
读取条码隔离模板信息
}
\examples{
barcode_ban_readFile()
}
