% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeBan.R
\name{barcode_ban_tpl}
\alias{barcode_ban_tpl}
\title{生成隔离条码模板}
\usage{
barcode_ban_tpl(conn = tsda::conn_rds("lcrds"))
}
\arguments{
\item{conn}{连接放在主数据库中}
}
\value{
返回列表数据
}
\description{
生成隔离条码模板
}
\examples{
barcode_ban_tpl()
}
