% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeBan.R
\name{barcode_ban_updateDB}
\alias{barcode_ban_updateDB}
\title{条码隔离更新数据}
\usage{
barcode_ban_updateDB(conn = tsda::conn_rds("LCERP"), data)
}
\arguments{
\item{conn}{连接}

\item{data}{数据}
}
\value{
无
}
\description{
条码隔离更新数据
}
\examples{
barcode_ban_updateDB()
}
