% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeBan.R
\name{barcode_banned_query}
\alias{barcode_banned_query}
\title{查询已经被禁用的条码}
\usage{
barcode_banned_query(conn = tsda::conn_rds("LCERP"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
查询已经被禁用的条码
}
\examples{
barcode_banned_query()
}
