% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_createTable.R
\name{barcode_sync_table}
\alias{barcode_sync_table}
\title{创建同步龙腾条码的表结构}
\usage{
barcode_sync_table(conn = tsda::conn_rds("LCERP"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
创建同步龙腾条码的表结构
}
\examples{
barcode_sync_table()
}
