% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intBarCode.R
\name{barcodeInner_sync_action}
\alias{barcodeInner_sync_action}
\title{针对条码功能进行同步}
\usage{
barcodeInner_sync_action(conn = tsda::conn_rds("LCERP"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
针对条码功能进行同步
}
\examples{
barcode_sync_action()
}
