% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bom_getKeyNoType}
\alias{bom_getKeyNoType}
\title{获取件号的类型}
\usage{
bom_getKeyNoType(
  conn = tsda::conn_rds("lcrds"),
  FchartNo = "YX200A714",
  keyNo = ""
)
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{图号}

\item{keyNo}{件号}
}
\value{
返回值
}
\description{
获取件号的类型
}
\examples{
bom_getKeyNoType()
}
