% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bom_getVarValueType}
\alias{bom_getVarValueType}
\title{针对从L表中的取出来的数值进行判断}
\usage{
bom_getVarValueType(value)
}
\arguments{
\item{value}{数值}
}
\value{
返回值
}
\description{
针对从L表中的取出来的数值进行判断
}
\examples{
bom_getVarValueType()
}
