% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom_split.R
\name{bom_split}
\alias{bom_split}
\title{针对BOM数据进行拆分处理}
\usage{
bom_split(
  mtrl_multiple_G = c("P203031A112G08G02G01G21G34G10G06G44G31",
    "P203031A112G08G02G01G21G34G10G06G44G32"),
  maxId = 0
)
}
\arguments{
\item{mtrl_multiple_G}{待处理的物料包含多个G}

\item{maxId}{最大的内码号}
}
\value{
返回一个处理后的数据框
}
\description{
针对BOM数据进行拆分处理
}
\examples{
bom_split()
}
