% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom_split.R
\name{bom_split_getMaxId}
\alias{bom_split_getMaxId}
\title{获取BOM拆分的最大号}
\usage{
bom_split_getMaxId(conn = tsda::conn_rds("lcrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
获取BOM拆分的最大号
}
\examples{
bom_split_getMaxId()
}
