% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom_split.R
\name{bom_split_query}
\alias{bom_split_query}
\title{bom数据查询}
\usage{
bom_split_query(conn = tsda::conn_rds("lcrds"), FBillNo = "")
}
\arguments{
\item{conn}{连接}

\item{FBillNo}{单据编号}
}
\value{
返回值
}
\description{
bom数据查询
}
\examples{
bom_split_query()
}
