% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom_split.R
\name{bom_split_upload}
\alias{bom_split_upload}
\title{BOM拆分删除数据}
\usage{
bom_split_upload(conn = tsda::conn_rds("lcrds"), data = bom_split())
}
\arguments{
\item{conn}{连接}

\item{data}{数据}
}
\value{
无返回值
}
\description{
BOM拆分删除数据
}
\examples{
bom_split_upload_input()
}
