% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{db_bom_setUpdate}
\alias{db_bom_setUpdate}
\title{针对上传的图号完成处理后设置标志为1}
\usage{
db_bom_setUpdate(conn = tsda::conn_rds("lcrds"), FchartNo = "SYE601B672")
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}
}
\value{
无返回值
}
\description{
针对上传的图号完成处理后设置标志为1
}
\examples{
db_bom_setUpdate()
}
