% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dm_bom_deleteDB}
\alias{dm_bom_deleteDB}
\title{针对数据进行处理}
\usage{
dm_bom_deleteDB(
  conn = tsda::conn_rds("lcrds"),
  FchartNo = "SYE601B672",
  FGtab = "G10",
  FLtab = "L02"
)
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}

\item{FGtab}{G}

\item{FLtab}{L}
}
\value{
return value
}
\description{
针对数据进行处理
}
\examples{
dm_bom_deleteDB()
}
