% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom.R
\name{dm_chartNo_delete_byGL}
\alias{dm_chartNo_delete_byGL}
\title{按主图号删除数据库,按番与L番进行处理}
\usage{
dm_chartNo_delete_byGL(
  conn = tsda::conn_rds("lcrds"),
  FchartNo = "SYE601B672",
  FParamG = "G01",
  FParamL = "L57"
)
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}

\item{FParamG}{G番}

\item{FParamL}{L番}
}
\value{
返回值
}
\description{
按主图号删除数据库,按番与L番进行处理
}
\examples{
dm_chartNo_delete_byGL()
}
