% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom.R
\name{dm_chartNo_deleteDB}
\alias{dm_chartNo_deleteDB}
\title{按主图号删除数据库}
\usage{
dm_chartNo_deleteDB(conn = tsda::conn_rds("lcrds"), FchartNo = "SYE601B672")
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}
}
\value{
返回值
}
\description{
按主图号删除数据库
}
\examples{
dm_chartNo_deleteDB()
}
