% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dm_dealAll}
\alias{dm_dealAll}
\title{更新所有BOM}
\usage{
dm_dealAll(conn = tsda::conn_rds("lcrds"), show_process = FALSE)
}
\arguments{
\item{conn}{连接}

\item{show_process}{是否显示进度}
}
\value{
返回值
}
\description{
更新所有BOM
}
\examples{
dm_dealAll()
}
