% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dm_dealAll2}
\alias{dm_dealAll2}
\title{写入数据库}
\usage{
dm_dealAll2(conn = tsda::conn_rds("lcrds"), show_process = FALSE)
}
\arguments{
\item{conn}{连接}

\item{show_process}{是否显示进度}
}
\value{
返回值
}
\description{
写入数据库
}
\examples{
dm_dealAll2()
}
