% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dm_dealOne}
\alias{dm_dealOne}
\title{处理单个dm清单}
\usage{
dm_dealOne(
  conn = tsda::conn_rds("lcrds"),
  FchartNo = "SYE601B672",
  FGtab = "G10",
  FLtab = "L02"
)
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}

\item{FGtab}{G表}

\item{FLtab}{L表值}
}
\value{
返回值
}
\description{
处理单个dm清单
}
\examples{
dm_dealOne()
}
