% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dm_getToDoChartNo}
\alias{dm_getToDoChartNo}
\title{获取待处理的图号}
\usage{
dm_getToDoChartNo(conn = tsda::conn_rds("lcrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回主图事情
}
\description{
获取待处理的图号
}
\examples{
dm_getToDoChartNo()
}
