% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dm_queryAll}
\alias{dm_queryAll}
\title{查询所有清单}
\usage{
dm_queryAll(
  file = "data-raw/bom_src4.xlsx",
  sheet = "DM清单",
  conn = tsda::conn_rds("lcrds")
)
}
\arguments{
\item{file}{文件}

\item{sheet}{页答}

\item{conn}{连接}
}
\value{
返回值
}
\description{
查询所有清单
}
\examples{
dm_queryAll()
}
