% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dm_queryAll2}
\alias{dm_queryAll2}
\title{查询DM清单数据}
\usage{
dm_queryAll2(
  file = "data-raw/bom_src4.xlsx",
  sheet = "DM清单",
  conn = tsda::conn_rds("lcrds")
)
}
\arguments{
\item{file}{文件}

\item{sheet}{页答}

\item{conn}{连接}
}
\value{
返回值
}
\description{
查询DM清单数据
}
\examples{
dm_queryAll2()
}
