% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dm_queryAll_writeDB}
\alias{dm_queryAll_writeDB}
\title{写入数据库}
\usage{
dm_queryAll_writeDB(
  file = "data-raw/bom_src4.xlsx",
  sheet = "DM清单",
  conn = tsda::conn_rds("lcrds")
)
}
\arguments{
\item{file}{文件}

\item{sheet}{页签}

\item{conn}{连接}
}
\value{
返回值
}
\description{
写入数据库
}
\examples{
dm_queryAll_writeDB()
}
