% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dm_read_list}
\alias{dm_read_list}
\title{读取DM清单}
\usage{
dm_read_list(file = "data-raw/bom_src4.xlsx", sheet = "DM清单")
}
\arguments{
\item{file}{文件}

\item{sheet}{页答名}
}
\value{
返回值
}
\description{
读取DM清单
}
\examples{
dm_read_list()
}
