% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dm_ReadBy_ChartNo}
\alias{dm_ReadBy_ChartNo}
\title{按图号读取数据}
\usage{
dm_ReadBy_ChartNo(conn = tsda::conn_rds("lcrds"), FchartNo = "SE304A200")
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}
}
\value{
返回值
}
\description{
按图号读取数据
}
\examples{
dm_ReadBy_ChartNo()
}
