% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataStat.R
\name{dm_ReadBy_ChartNo_GL_dealOne}
\alias{dm_ReadBy_ChartNo_GL_dealOne}
\title{获取相当数据,处理每一条记录}
\usage{
dm_ReadBy_ChartNo_GL_dealOne(
  conn = tsda::conn_rds("lcrds"),
  FchartNo = "SE304A200",
  FParamG = "G01",
  FParamL = "L33",
  page_size = 300
)
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{图号}

\item{FParamG}{参数}

\item{FParamL}{参数}

\item{page_size}{每页大小}
}
\value{
返回值
}
\description{
获取相当数据,处理每一条记录
}
\examples{
dm_ReadBy_ChartNo_GL_dealOne()
}
