% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom.R
\name{dm_ReadBy_ChartNo_Ltab}
\alias{dm_ReadBy_ChartNo_Ltab}
\title{更新记录}
\usage{
dm_ReadBy_ChartNo_Ltab(
  conn = tsda::conn_rds("lcrds"),
  FchartNo = "YX200A714",
  FLtab = "L07"
)
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}

\item{FLtab}{L番}
}
\value{
返回值
}
\description{
更新记录
}
\examples{
dm_ReadBy_ChartNo_Ltab()
}
