% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataStat.R
\name{dm_ReadBy_ChartNo_LtabBatch_dealBom}
\alias{dm_ReadBy_ChartNo_LtabBatch_dealBom}
\title{针对数据进行处理}
\usage{
dm_ReadBy_ChartNo_LtabBatch_dealBom(
  conn = tsda::conn_rds("lcrds"),
  FchartNo = "SE304A200",
  FParamG = "G01",
  FParamL = "L33",
  page_size = 300
)
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{图号}

\item{page_size}{页面大小}

\item{FGtab}{G番}

\item{FLtabBatch}{L番或多个}
}
\value{
返回值
}
\description{
针对数据进行处理
}
\examples{
dm_ReadBy_ChartNo_LtabBatch_dealBom()
}
