% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dm_reCalcChartNo_withoutLtab}
\alias{dm_reCalcChartNo_withoutLtab}
\title{重算只有G番没有L番的历史数据}
\usage{
dm_reCalcChartNo_withoutLtab(conn = tsda::conn_rds("lcrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
重算只有G番没有L番的历史数据
}
\examples{
dm_reCalcChartNo_withoutLtab()
}
