% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dm_selectDB_detail}
\alias{dm_selectDB_detail}
\title{查询信息}
\usage{
dm_selectDB_detail(
  conn = tsda::conn_rds("lcrds"),
  FchartNo = "P235009B198",
  FParamG = "G11",
  FParamL = "L42"
)
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}

\item{FParamG}{G番}

\item{FParamL}{L番}
}
\value{
返回值
}
\description{
查询信息
}
\examples{
dm_selectDB_detail()
}
