% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom.R
\name{dm_selectDB_detail2}
\alias{dm_selectDB_detail2}
\title{BOM查询多L番}
\usage{
dm_selectDB_detail2(
  conn = tsda::conn_rds("lcrds"),
  FchartNo = "SE304A200",
  FParamG = "G01",
  FParamL = ""
)
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}

\item{FParamG}{G番}

\item{FParamL}{L番}
}
\value{
返回值
}
\description{
BOM查询多L番
}
\examples{
dm_selectDB_detail2()
}
