% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom.R
\name{dm_selectDB_detail_combo1}
\alias{dm_selectDB_detail_combo1}
\title{针对图号进行查询，兼容DM清单结果}
\usage{
dm_selectDB_detail_combo1(
  conn = tsda::conn_rds("lcrds"),
  FchartNo = "YX200A714",
  FParamG = "GS11",
  FParamL = ""
)
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{图号}

\item{FParamG}{G番}

\item{FParamL}{L番}
}
\value{
返回值
}
\description{
针对图号进行查询，兼容DM清单结果
}
\examples{
dm_selectDB_detail_combo1()
}
