% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom.R
\name{dm_selectDB_query}
\alias{dm_selectDB_query}
\title{从数据库中直接查询BOM}
\usage{
dm_selectDB_query(
  conn = tsda::conn_rds("lcrds"),
  FchartNo = "SE304A200",
  FParamG = "G01",
  FParamL = ""
)
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}

\item{FParamG}{G番}

\item{FParamL}{L番}
}
\value{
返回值
}
\description{
从数据库中直接查询BOM
}
\examples{
dm_selectDB_query()
}
