% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dm_writeDB_ChartNo}
\alias{dm_writeDB_ChartNo}
\title{将图号数据写入数据库}
\usage{
dm_writeDB_ChartNo(conn = tsda::conn_rds("lcrds"), FchartNo = "SYE601B672")
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}
}
\value{
返回值
}
\description{
将图号数据写入数据库
}
\examples{
dm_writeDB_ChartNo()
}
