% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dm_writeDB_ChartNo_G}
\alias{dm_writeDB_ChartNo_G}
\title{将图号数据写入数据库,只更新指定G番}
\usage{
dm_writeDB_ChartNo_G(
  conn = tsda::conn_rds("lcrds"),
  FchartNo = "SE304A200",
  FParamG = "G01"
)
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}

\item{FParamG}{G番号}
}
\value{
返回值
}
\description{
将图号数据写入数据库,只更新指定G番
}
\examples{
dm_writeDB_ChartNo_G()
}
