% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dmList_createView}
\alias{dmList_createView}
\title{创建视图}
\usage{
dmList_createView(dms_token = "048017E3-CA7E-4DC7-BC87-0EA7D8C69C13")
}
\arguments{
\item{dms_token}{中台口令}
}
\value{
返回值
}
\description{
创建视图
}
\examples{
dmList_createView()
}
