% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dmList_data_sync}
\alias{dmList_data_sync}
\title{针对数据进行处理}
\usage{
dmList_data_sync(conn = tsda::conn_rds("lcrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
针对数据进行处理
}
\examples{
dmList_data_sync()
}
