% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dmList_Expand_Multi}
\alias{dmList_Expand_Multi}
\title{针对上传的DM清单进行多级展开的入口函数}
\usage{
dmList_Expand_Multi(
  file = "data-raw/bom_src4.xlsx",
  sheet = "DM清单",
  conn = tsda::conn_rds("lcrds")
)
}
\arguments{
\item{file}{文件}

\item{sheet}{页签}

\item{conn}{连接}
}
\value{
返回值
}
\description{
针对上传的DM清单进行多级展开的入口函数
}
\examples{
dmList_Expand_Multi()
}
