% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceModel.R
\name{dmList_getBillNo2}
\alias{dmList_getBillNo2}
\title{获取DM单的所有的单据编号}
\usage{
dmList_getBillNo2(dms_token = "048017E3-CA7E-4DC7-BC87-0EA7D8C69C13")
}
\arguments{
\item{dms_token}{口令}
}
\value{
返回值
}
\description{
获取DM单的所有的单据编号
}
\examples{
dmList_getBillNo()
}
