% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceModel.R
\name{dmList_LevelMax}
\alias{dmList_LevelMax}
\title{获取DM最大的级次}
\usage{
dmList_LevelMax(
  dms_token = "048017E3-CA7E-4DC7-BC87-0EA7D8C69C13",
  FDmNo = "DMP235000B1561920"
)
}
\arguments{
\item{dms_token}{口令}

\item{FDmNo}{DM单号}
}
\value{
返回值
}
\description{
获取DM最大的级次
}
\examples{
dmList_LevelMax()
}
