% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceModel.R
\name{dmList_LevelParentNo}
\alias{dmList_LevelParentNo}
\title{获取的某一级次的行号清单}
\usage{
dmList_LevelParentNo(
  dms_token = "048017E3-CA7E-4DC7-BC87-0EA7D8C69C13",
  FDmNo = "DMP235000B1561920",
  FLevel = 2
)
}
\arguments{
\item{dms_token}{口令}

\item{FDmNo}{DM单号}
}
\value{
返回值
}
\description{
获取的某一级次的行号清单
}
\examples{
dmList_LevelParentNo()
}
