% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dmList_need_toDo}
\alias{dmList_need_toDo}
\title{判断是否需要处理}
\usage{
dmList_need_toDo(conn = tsda::conn_rds("lcrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
判断是否需要处理
}
\examples{
dmList_need_toDo()
}
