% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dmList_queryAll_Multi}
\alias{dmList_queryAll_Multi}
\title{多次处理相关数据}
\usage{
dmList_queryAll_Multi(conn = tsda::conn_rds("lcrds"), data)
}
\arguments{
\item{conn}{连接}

\item{data}{处理数据}
}
\value{
返回值
}
\description{
多次处理相关数据
}
\examples{
dmList_queryAll_Multi()
}
