% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dmList_readDB_Input}
\alias{dmList_readDB_Input}
\title{读取DM清单数据}
\usage{
dmList_readDB_Input(conn = tsda::conn_rds("lcrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
读取DM清单数据
}
\examples{
dmList_readDB_Input()
}
