% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dmList_readDB_Input_cn}
\alias{dmList_readDB_Input_cn}
\title{读取DM数据,提供中文名}
\usage{
dmList_readDB_Input_cn(conn = tsda::conn_rds("lcrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
读取DM数据,提供中文名
}
\examples{
dmList_readDB_Input_cn()
}
