% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dmList_toDo_Multi}
\alias{dmList_toDo_Multi}
\title{多次读取待处理的数据}
\usage{
dmList_toDo_Multi(conn = tsda::conn_rds("lcrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
多次读取待处理的数据
}
\examples{
dmList_toDo_Multi()
}
