% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceModel.R
\name{dmList_toODSBatch}
\alias{dmList_toODSBatch}
\title{查看清单的数据情况}
\usage{
dmList_toODSBatch(dms_token = "048017E3-CA7E-4DC7-BC87-0EA7D8C69C13")
}
\arguments{
\item{dms_token}{中台口令}
}
\value{
返回值
}
\description{
查看清单的数据情况
}
\examples{
dmList_toODSBatch()
}
