% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dmList_updateSubRowNo_Multi}
\alias{dmList_updateSubRowNo_Multi}
\title{更新序号及下级代码，方便后续进行排序查询}
\usage{
dmList_updateSubRowNo_Multi(conn = tsda::conn_rds("lcrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
更新序号及下级代码，方便后续进行排序查询
}
\examples{
dmList_updateSubRowNo_Multi()
}
