% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceModel.R
\name{dmListOds_createTable}
\alias{dmListOds_createTable}
\title{创建DM单ODS表中创建表结构}
\usage{
dmListOds_createTable(dms_token = "048017E3-CA7E-4DC7-BC87-0EA7D8C69C13")
}
\arguments{
\item{dms_token}{口令}
}
\value{
返回值
}
\description{
创建DM单ODS表中创建表结构
}
\examples{
dmListOds_createTable()
}
