% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceModel.R
\name{dmListODS_isNew}
\alias{dmListODS_isNew}
\title{判断是否存在新的DM单}
\usage{
dmListODS_isNew(
  dms_token = "048017E3-CA7E-4DC7-BC87-0EA7D8C69C13",
  FDmNo = "DMP235000B1561502"
)
}
\arguments{
\item{dms_token}{口令}

\item{FDmNo}{DM单}
}
\value{
返回值
}
\description{
判断是否存在新的DM单
}
\examples{
dmListODS_isNew()
}
