% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dmQuery1_readDB}
\alias{dmQuery1_readDB}
\title{提供DM清单的单个查询}
\usage{
dmQuery1_readDB(conn = tsda::conn_rds("lcrds"), FDmNo = "DMP235000B1561920")
}
\arguments{
\item{conn}{连接}

\item{FDmNo}{DM单号}
}
\value{
返回值
}
\description{
提供DM清单的单个查询
}
\examples{
dmQuery1_readDB()
}
