% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dmQuery1_readDB_cn}
\alias{dmQuery1_readDB_cn}
\title{提供对DM清单的单个查询并返回中文名列表}
\usage{
dmQuery1_readDB_cn(conn = tsda::conn_rds("lcrds"), FDmNo = "DMP235000B1561920")
}
\arguments{
\item{conn}{连接}

\item{FDmNo}{单个DM清单}
}
\value{
返回值
}
\description{
提供对DM清单的单个查询并返回中文名列表
}
\examples{
dmQuery1_readDB_cn()
}
