% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{dmQuery_Batch_file}
\alias{dmQuery_Batch_file}
\title{针对DM清单及BOM进行混合查询}
\usage{
dmQuery_Batch_file(
  file = "data-raw/DM配件混合查询.xlsx",
  sheet = "DM配件混合查询",
  conn = tsda::conn_rds("lcrds")
)
}
\arguments{
\item{file}{文件}

\item{sheet}{页签}

\item{conn}{连接}
}
\value{
返回值
}
\description{
针对DM清单及BOM进行混合查询
}
\examples{
dmQuery_Batch_file()
}
