% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extBarcode.R
\name{exrBarcode_getUnAllocated}
\alias{exrBarcode_getUnAllocated}
\title{查询未分配记录}
\usage{
exrBarcode_getUnAllocated(
  conn = tsda::conn_rds("lcrds"),
  FCalcNo = 11,
  FChartNo = "P207012C134G01",
  FNote_All = "G116*G103,G116,G107*G110,G363*G168,",
  n = 4
)
}
\arguments{
\item{conn}{连接}

\item{FCalcNo}{计算次数}

\item{FChartNo}{图号}

\item{FNote_All}{备注}

\item{n}{数量}
}
\value{
返回值
}
\description{
查询未分配记录
}
\examples{
exrBarcode_getUnAllocated()
}
