% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_allocation.R
\name{extBarcode_Allocate_ByEachNote}
\alias{extBarcode_Allocate_ByEachNote}
\title{针对数据填写分配逻辑}
\usage{
extBarcode_Allocate_ByEachNote(
  conn_rds = tsda::conn_rds("lcrds"),
  conn_erp = tsda::conn_rds("LCERP2"),
  FCalcNo = 11,
  FChartNo = "P207012C134G01",
  FNote_All = "G116*G103,G116,G107*G110,G363*G168,",
  n = 4
)
}
\arguments{
\item{conn_rds}{连接1}

\item{conn_erp}{连接2}

\item{FCalcNo}{计算号}

\item{FChartNo}{图号}

\item{FNote_All}{备注}

\item{n}{行数}
}
\value{
返回值
}
\description{
针对数据填写分配逻辑
}
\examples{
extBacode_Allocate_ByEachNote()
}
