% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_allocation.R
\name{extBarcode_AllocateALL}
\alias{extBarcode_AllocateALL}
\title{条码进行自动匹配}
\usage{
extBarcode_AllocateALL(
  conn_rds = tsda::conn_rds("lcrds"),
  conn_erp = tsda::conn_rds("LCERP")
)
}
\arguments{
\item{conn_rds}{链接1}

\item{conn_erp}{链接2}
}
\value{
返回值
}
\description{
条码进行自动匹配
}
\examples{
extBarcode_AllocateALL()
}
