% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_allocation.R
\name{extBarcode_AllocateResult}
\alias{extBarcode_AllocateResult}
\title{获取条码分配结果}
\usage{
extBarcode_AllocateResult(
  conn_rds = tsda::conn_rds("lcrds"),
  conn_erp = tsda::conn_rds("LCERP")
)
}
\arguments{
\item{conn_rds}{连接1}

\item{conn_erp}{连接2}
}
\value{
返回值
}
\description{
获取条码分配结果
}
\examples{
extBarcode_AllocateResult()
}
