% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extBarcode.R
\name{extBarcode_getNoteCount_currentCalc}
\alias{extBarcode_getNoteCount_currentCalc}
\title{获取当前计算号的信息}
\usage{
extBarcode_getNoteCount_currentCalc(conn = tsda::conn_rds("lcrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
获取当前计算号的信息
}
\examples{
extBarcode_getNoteCount_currentCalc()
}
