% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extBarcode.R
\name{extBarcode_MaxCalcNo}
\alias{extBarcode_MaxCalcNo}
\title{获取当前最大运行号}
\usage{
extBarcode_MaxCalcNo(conn = tsda::conn_rds("lcrds"))
}
\arguments{
\item{conn}{连接信息}
}
\value{
返回值
}
\description{
获取当前最大运行号
}
\examples{
extBarcode_MaxCalcNo()
}
