% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extBarcode.R
\name{extBarcode_newId}
\alias{extBarcode_newId}
\title{获取最大运算号}
\usage{
extBarcode_newId(conn = tsda::conn_rds("lcrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返顺值
}
\description{
获取最大运算号
}
\examples{
extBarcode_newId()
}
