% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extBarcode.R
\name{extBarcode_Query_CurrentCalcNo}
\alias{extBarcode_Query_CurrentCalcNo}
\title{获取当前运算单号的外部条件}
\usage{
extBarcode_Query_CurrentCalcNo(conn = tsda::conn_rds("lcrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回结果
}
\description{
获取当前运算单号的外部条件
}
\examples{
extBarcode_Query_CurrentCalcNo()
}
