% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extBarcode.R
\name{extBarcode_query_mfgNote}
\alias{extBarcode_query_mfgNote}
\title{生产订单备注}
\usage{
extBarcode_query_mfgNote(conn = tsda::conn_rds("lcrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
生产订单备注
}
\examples{
extcode_query_mfgNote()
}
