% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extBarcode.R
\name{extBarcode_query_NoteALL}
\alias{extBarcode_query_NoteALL}
\title{获取当前运算号的所有备注信息}
\usage{
extBarcode_query_NoteALL(conn = tsda::conn_rds("lcrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
获取当前运算号的所有备注信息
}
\examples{
extBarcode_query_NoteALL()
}
