% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extBarcode.R
\name{extBarcode_query_TechNote}
\alias{extBarcode_query_TechNote}
\title{外部订单查询技术备注}
\usage{
extBarcode_query_TechNote(conn = tsda::conn_rds("lcrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回查询结果
}
\description{
外部订单查询技术备注
}
\examples{
extcode_query_TechNote()
}
