% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extBarcode.R
\name{extBarcode_QueryByCalcNo}
\alias{extBarcode_QueryByCalcNo}
\title{按指定运算号查询外部条件}
\usage{
extBarcode_QueryByCalcNo(conn = tsda::conn_rds("lcrds"), calcNo = 11)
}
\arguments{
\item{conn}{连接}

\item{calcNo}{运算号}
}
\value{
返回值
}
\description{
按指定运算号查询外部条件
}
\examples{
extBarcode_QueryByCalcNo()
}
