% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extBarcode.R
\name{extBarcode_read}
\alias{extBarcode_read}
\title{读取外部条码信息}
\usage{
extBarcode_read(
  file = "data-raw/外部订单模板.xlsx",
  FCalcNo = "LSD001",
  lang = "cn"
)
}
\arguments{
\item{file}{文件}
}
\value{
返回值
}
\description{
读取外部条码信息
}
\examples{
extBarcode_read()
}
