% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtable.R
\name{Gtab_delete_db}
\alias{Gtab_delete_db}
\title{G表删除}
\usage{
Gtab_delete_db(conn = tsda::conn_rds("lcrds"), FchartNo = "P235067C156")
}
\arguments{
\item{conn}{逻辑}

\item{FchartNo}{主图号}
}
\value{
返回值
}
\description{
G表删除
}
\examples{
Gtab_delete_db()
}
