% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtable.R
\name{Gtab_get_uniqueMembers}
\alias{Gtab_get_uniqueMembers}
\title{获取G表中的唯一元素}
\usage{
Gtab_get_uniqueMembers(conn = tsda::conn_rds("lcrds"), FchartNo = "YE603A049")
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}
}
\value{
返回值
}
\description{
获取G表中的唯一元素
}
\examples{
Gtab_get_uniqueMembers()
}
