% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtable.R
\name{Gtab_is_new}
\alias{Gtab_is_new}
\title{判断G表图号是否为新的内容}
\usage{
Gtab_is_new(conn = tsda::conn_rds("lcrds"), FchartNo = "P235067C156")
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}
}
\value{
返回值
}
\description{
判断G表图号是否为新的内容
}
\examples{
Gtab_is_new()
}
