% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtable.R
\name{Gtab_read}
\alias{Gtab_read}
\title{读取G番表}
\usage{
Gtab_read(
  file = "./data-raw/bom_src.xlsx",
  sheetName = "P235911B000",
  max_row = 198,
  rm.na = TRUE
)
}
\arguments{
\item{file}{文件}

\item{sheetName}{页答名称}

\item{max_row}{最大行数，默认99}
}
\value{
返回值
}
\description{
读取G番表
}
\examples{
lc_read_G_table()
}
