% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtable.R
\name{Gtab_select_db}
\alias{Gtab_select_db}
\title{从数据库中读取G表}
\usage{
Gtab_select_db(
  conn = tsda::conn_rds("lcrds"),
  FchartNo = "P235067C156",
  FGtab = "G01"
)
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}

\item{FGtab}{G番号}
}
\value{
返回值
}
\description{
从数据库中读取G表
}
\examples{
Gtab_select_db()
}
