% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtable.R
\name{Gtab_select_db_calc}
\alias{Gtab_select_db_calc}
\title{从数据库中读取G表,用于计算BOM}
\usage{
Gtab_select_db_calc(
  conn = tsda::conn_rds("lcrds"),
  FchartNo = "SE304A200",
  FGtab = "G01"
)
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}

\item{FGtab}{G番号}
}
\value{
返回值
}
\description{
从数据库中读取G表,用于计算BOM
}
\examples{
Gtab_select_db_calc()
}
