% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtable.R
\name{Gtab_selectDB_byChartNo}
\alias{Gtab_selectDB_byChartNo}
\title{从数据库中读取G表}
\usage{
Gtab_selectDB_byChartNo(
  conn = tsda::conn_rds("lcrds"),
  FchartNo = "P235067C156"
)
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}
}
\value{
返回值
}
\description{
从数据库中读取G表
}
\examples{
Gtab_select_db()
}
