% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtable.R
\name{Gtab_selectDB_byChartNo2}
\alias{Gtab_selectDB_byChartNo2}
\title{查询相关}
\usage{
Gtab_selectDB_byChartNo2(
  conn = tsda::conn_rds("lcrds"),
  FchartNo = "YE603A049"
)
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}
}
\value{
返回值
}
\description{
查询相关
}
\examples{
Gtab_selectDB_byChartNo2()
}
