% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtable.R
\name{Gtab_standard}
\alias{Gtab_standard}
\title{针对G番进行标准化处理,保留真正的L番，其他值设置为空}
\usage{
Gtab_standard(x)
}
\arguments{
\item{x}{数据}
}
\value{
返回值
}
\description{
针对G番进行标准化处理,保留真正的L番，其他值设置为空
}
\examples{
Gtab_standard()
}
