% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtable.R
\name{Gtab_write_db}
\alias{Gtab_write_db}
\title{G表写入}
\usage{
Gtab_write_db(
  conn = tsda::conn_rds("lcrds"),
  file = "./data-raw/bom_src.xlsx",
  sheetName = "P235911B000",
  max_row = 98,
  rm.na = TRUE,
  is_force
)
}
\arguments{
\item{conn}{连接}

\item{file}{文件}

\item{sheetName}{页签}

\item{max_row}{最大行}

\item{rm.na}{删除空格}
}
\value{
返回值
}
\description{
G表写入
}
\examples{
Gtab_write_db()
}
