% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtable.R
\name{Gtab_write_db_stat}
\alias{Gtab_write_db_stat}
\title{从数据库中读取G表,用于计算G表的统计信息,写入数据库}
\usage{
Gtab_write_db_stat(
  conn = tsda::conn_rds("lcrds"),
  FchartNo = "SE304A200",
  FGtab = "G01",
  page_size = 200
)
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}

\item{FGtab}{G番号}

\item{page_size}{分批写入数据}
}
\value{
返回值
}
\description{
从数据库中读取G表,用于计算G表的统计信息,写入数据库
}
\examples{
Gtab_select_db_stat()
}
