% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_getDetail}
\alias{log_getDetail}
\title{查看最近日志记录}
\usage{
log_getDetail(conn_erp = tsda::conn_rds("LCERP2"), topN = 1)
}
\arguments{
\item{conn_erp}{ERP}

\item{topN}{第N次}
}
\value{
返回值
}
\description{
查看最近日志记录
}
\examples{
返回值
log_getCurrent()
}
