% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ltable.R
\name{Ltab_batchWrite_db}
\alias{Ltab_batchWrite_db}
\title{批量写入L番表进入数据库}
\usage{
Ltab_batchWrite_db(
  conn = tsda::conn_rds("lcrds"),
  file = "./data-raw/bom_src.xlsx",
  exclude_sheetName = "YE604B797",
  show_progress = FALSE,
  skip_rows = 98,
  rm.na = TRUE
)
}
\arguments{
\item{conn}{连接}

\item{file}{文件}

\item{exclude_sheetName}{排除页签}

\item{show_progress}{是否显示进度}

\item{skip_rows}{跳过行数}

\item{rm.na}{删除空行}
}
\value{
返回值
}
\description{
批量写入L番表进入数据库
}
\examples{
Ltab_batchWrite_db()
}
