% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ltable.R
\name{Ltab_checkExist}
\alias{Ltab_checkExist}
\title{检查L番是否存在}
\usage{
Ltab_checkExist(conn = tsda::conn_rds("lcrds"), FchartNo = "YE603A049")
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}
}
\value{
返回值
}
\description{
检查L番是否存在
}
\examples{
Ltab_checkExist()
}
