% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ltable.R
\name{Ltab_delete_db}
\alias{Ltab_delete_db}
\title{L表删除}
\usage{
Ltab_delete_db(conn = tsda::conn_rds("lcrds"), FchartNo = "P235067C156")
}
\arguments{
\item{conn}{逻辑}

\item{FchartNo}{主图号}
}
\value{
返回值
}
\description{
L表删除
}
\examples{
Ltab_delete_db()
}
