% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ltable.R
\name{Ltab_get_uniqueMembers}
\alias{Ltab_get_uniqueMembers}
\title{获取L表不重复元素}
\usage{
Ltab_get_uniqueMembers(conn = tsda::conn_rds("lcrds"), FchartNo = "YE603A049")
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}
}
\value{

}
\description{
获取L表不重复元素
}
\examples{
Ltab_get_uniqueMembers()
}
