% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ltable.R
\name{Ltab_get_uniqueVars}
\alias{Ltab_get_uniqueVars}
\title{L表获取唯一变量}
\usage{
Ltab_get_uniqueVars(conn = tsda::conn_rds("lcrds"), FchartNo = "YE603A049")
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}
}
\value{
返回唯一变量
}
\description{
L表获取唯一变量
}
\examples{
Ltab_get_uniqueVars()
}
