% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ltable.R
\name{Ltab_get_varValue}
\alias{Ltab_get_varValue}
\title{获取L表的变量值}
\usage{
Ltab_get_varValue(
  conn = tsda::conn_rds("lcrds"),
  FchartNo = "SYE601B672",
  FLtab = "L02",
  FkeyNo = "A1L"
)
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}

\item{FLtab}{L番值}

\item{FkeyNo}{变量值}
}
\value{
返回具有数值
}
\description{
获取L表的变量值
}
\examples{
Ltab_get_varValue()
}
