% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ltable.R
\name{Ltab_read}
\alias{Ltab_read}
\title{读取L番表}
\usage{
Ltab_read(
  file = "./data-raw/bom_src.xlsx",
  sheetName = "YE604B797",
  skip_rows = 98,
  rm.na = TRUE
)
}
\arguments{
\item{file}{文件名称xlsx}

\item{sheetName}{页答名称}

\item{skip_rows}{一般从第100行开始，跳过98行}

\item{rm.na是否去除}{}
}
\value{
返回值
}
\description{
读取L番表
}
\examples{
Ltab_read()
}
