% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ltable.R
\name{Ltab_select_db}
\alias{Ltab_select_db}
\title{L翻表}
\usage{
Ltab_select_db(conn = tsda::conn_rds("lcrds"), FchartNo = "YE603A049")
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}
}
\value{
返回值
}
\description{
L翻表
}
\examples{
Ltab_select_db()
}
