% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ltable.R
\name{Ltab_select_db2}
\alias{Ltab_select_db2}
\title{查询L表相关信息}
\usage{
Ltab_select_db2(conn = tsda::conn_rds("lcrds"), FchartNo = "YE603A049")
}
\arguments{
\item{conn}{连接}

\item{FchartNo}{主图号}
}
\value{
返回值
}
\description{
查询L表相关信息
}
\examples{
Ltab_select_db2()
}
