% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ltable.R
\name{Ltab_standard}
\alias{Ltab_standard}
\title{针对L番进行标准化处理,保留真正的L番，其他值设置为空}
\usage{
Ltab_standard(x)
}
\arguments{
\item{x}{数据}
}
\value{
返回值
}
\description{
针对L番进行标准化处理,保留真正的L番，其他值设置为空
}
\examples{
Ltab_standard()
}
