% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ltable.R
\name{Ltab_write_db}
\alias{Ltab_write_db}
\title{写入数据库}
\usage{
Ltab_write_db(
  conn = tsda::conn_rds("lcrds"),
  file = "./data-raw/bom_src.xlsx",
  sheetName = "YE604B797",
  skip_rows = 199,
  rm.na = TRUE
)
}
\arguments{
\item{conn}{连接}

\item{file}{文件}

\item{sheetName}{页签名}

\item{skip_rows}{行数}

\item{rm.na}{去除空}
}
\value{
返回值
}
\description{
写入数据库
}
\examples{
Ltab_write_db()
}
