% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material.R
\name{material_ErpSyncDms}
\alias{material_ErpSyncDms}
\title{返回物料数据同步}
\usage{
material_ErpSyncDms(
  erp_token = "BF6CBEB5-BC3E-422B-B96C-32BDACCE32ED",
  dms_token = "048017E3-CA7E-4DC7-BC87-0EA7D8C69C13"
)
}
\arguments{
\item{erp_token}{erp口令}

\item{dms_token}{DMS口令}
}
\value{
返回值
}
\description{
返回物料数据同步
}
\examples{
material_ErpSyncDms()
}
