% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{page_setting}
\alias{page_setting}
\title{分页设置}
\usage{
page_setting(totalRow = 14, pageRow = 4)
}
\arguments{
\item{totalRow}{总行数}

\item{pageRow}{每页行数}
}
\value{
返回列表
}
\description{
分页设置
}
\examples{
page_setting()
}
