% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceDetail.R
\name{priceDetail_query}
\alias{priceDetail_query}
\title{查询核价明细表}
\usage{
priceDetail_query(
  dms_token = "048017E3-CA7E-4DC7-BC87-0EA7D8C69C13",
  FDmNo = "DMP235000B1561502"
)
}
\arguments{
\item{dms_token}{口令}

\item{FDmNo}{DM单}
}
\value{
返回值
}
\description{
查询核价明细表
}
\examples{
priceDetail_query()
}
