% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceModel.R
\name{priceModel_createTable}
\alias{priceModel_createTable}
\title{价格模型创建表结构}
\usage{
priceModel_createTable(dms_token = "048017E3-CA7E-4DC7-BC87-0EA7D8C69C13")
}
\arguments{
\item{dms_token}{中台口令}
}
\value{
返回值
}
\description{
价格模型创建表结构
}
\examples{
priceModel_createTable()
}
