% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceModel.R
\name{priceModel_DmCalcOne}
\alias{priceModel_DmCalcOne}
\title{处理一条记录}
\usage{
priceModel_DmCalcOne(
  dms_token = "048017E3-CA7E-4DC7-BC87-0EA7D8C69C13",
  FDmNo = "DMP235000B1561920"
)
}
\arguments{
\item{dms_token}{中台口令}

\item{FDmNo}{单号}
}
\value{
返回值
}
\description{
处理一条记录
}
\examples{
priceModel_DmCalcOne()
}
