% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceModel.R
\name{priceModel_DmErrorDel}
\alias{priceModel_DmErrorDel}
\title{DM异常处理，删除只有Level等于0的记录}
\usage{
priceModel_DmErrorDel(dms_token = "048017E3-CA7E-4DC7-BC87-0EA7D8C69C13")
}
\arguments{
\item{dms_token}{口令}
}
\value{
返回值
}
\description{
DM异常处理，删除只有Level等于0的记录
}
\examples{
priceModel_DmErrorDel()
}
