% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceModel.R
\name{priceModel_DmStatus}
\alias{priceModel_DmStatus}
\title{更新DM单状态}
\usage{
priceModel_DmStatus(
  dms_token = "048017E3-CA7E-4DC7-BC87-0EA7D8C69C13",
  FDmNo = "DMP235000B1561920"
)
}
\arguments{
\item{dms_token}{口令}

\item{FDmNo}{DM单号}
}
\value{
返回值
}
\description{
更新DM单状态
}
\examples{
priceModel_DmStatus()
}
