% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceModel.R
\name{priceModel_getPriceAll}
\alias{priceModel_getPriceAll}
\title{获取物料检查信息}
\usage{
priceModel_getPriceAll(
  dms_token = "048017E3-CA7E-4DC7-BC87-0EA7D8C69C13",
  FChartNumbers = c("X39DM-230", "X39DM-204", "X39BX-43", "X39BX-01", "Z148B0.30-WH-")
)
}
\arguments{
\item{dms_token}{口令}

\item{FChartNumbers}{图号}
}
\value{
返回值
}
\description{
获取物料检查信息
}
\examples{
priceModel_getPriceAll()
}
