% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceModel.R
\name{priceModel_getPriceOne}
\alias{priceModel_getPriceOne}
\title{获取物料检查信息}
\usage{
priceModel_getPriceOne(
  dms_token = "048017E3-CA7E-4DC7-BC87-0EA7D8C69C13",
  FChartNumber = "X39DM-230"
)
}
\arguments{
\item{dms_token}{口令}

\item{FChartNumber}{图号}
}
\value{
返回值
}
\description{
获取物料检查信息
}
\examples{
priceModel_getPriceOne()
}
