% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceModel.R
\name{priceModel_Unit}
\alias{priceModel_Unit}
\title{查询数据}
\usage{
priceModel_Unit(
  dms_token = "048017E3-CA7E-4DC7-BC87-0EA7D8C69C13",
  FDmNo = "DMP235000B1561920",
  FLevel = 2,
  FParentNo = "0080020"
)
}
\arguments{
\item{dms_token}{口令}

\item{FDmNo}{DM单号}

\item{FLevel}{级次}

\item{FParentNo}{行号}
}
\value{
返回值
}
\description{
查询数据
}
\examples{
priceModel_Unit()
}
