% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceSummary.R
\name{priceSummary_query}
\alias{priceSummary_query}
\title{查询核价明细表}
\usage{
priceSummary_query(dms_token = "048017E3-CA7E-4DC7-BC87-0EA7D8C69C13")
}
\arguments{
\item{dms_token}{口令}
}
\value{
返回值
}
\description{
查询核价明细表
}
\examples{
priceSummary_query()
}
