% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchasePrice.R
\name{purchasePriceDms_createTable}
\alias{purchasePriceDms_createTable}
\title{创建表结构}
\usage{
purchasePriceDms_createTable(
  dms_token = "048017E3-CA7E-4DC7-BC87-0EA7D8C69C13"
)
}
\arguments{
\item{dms_token}{DMS口令}
}
\value{
返回值
}
\description{
创建表结构
}
\examples{
purchasePriceLast_createTable()
}
