% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchasePrice.R
\name{purchasePriceErp_query}
\alias{purchasePriceErp_query}
\title{返回采购单价查询}
\usage{
purchasePriceErp_query(erp_token = "BF6CBEB5-BC3E-422B-B96C-32BDACCE32ED")
}
\arguments{
\item{erp_token}{ERP口令}
}
\value{
返回值
}
\description{
返回采购单价查询
}
\examples{
purchasePrice_query()
}
