% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchasePrice.R
\name{purchsePriceDms_query}
\alias{purchsePriceDms_query}
\title{采购价格最新数据查询}
\usage{
purchsePriceDms_query(
  dms_token = "048017E3-CA7E-4DC7-BC87-0EA7D8C69C13",
  FChartNumber = ""
)
}
\arguments{
\item{dms_token}{dms口令}

\item{FChartNumber}{图号}
}
\value{
返回值
}
\description{
采购价格最新数据查询
}
\examples{
purchsePriceLastRds_query()
}
