% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intBarCode.R
\name{query_barcode_chartNo}
\alias{query_barcode_chartNo}
\title{查询内部条码}
\usage{
query_barcode_chartNo(
  conn = tsda::conn_rds("lcrds"),
  fchartNo = "P207012C134G01"
)
}
\arguments{
\item{conn}{添加连接}

\item{fchartNo}{图号}
}
\value{
返回值
}
\description{
查询内部条码
}
\examples{
query_barcode_chartNo()
}
