% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartMtrlMap.R
\name{read_chartMtrlMapping}
\alias{read_chartMtrlMapping}
\title{读取图号及物料对照表}
\usage{
read_chartMtrlMapping(
  conn = tsda::conn_rds("lcrds"),
  file = "data-raw/图号物料匹配表.xlsx",
  sheet = "图号物料匹配表"
)
}
\arguments{
\item{conn}{连接}

\item{file}{文件}

\item{sheet}{页签}
}
\value{
返顺值
}
\description{
读取图号及物料对照表
}
\examples{
read_chartMtrlMapping()
}
