% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soNote.R
\name{soNote_flag}
\alias{soNote_flag}
\title{自动判断相应的数据}
\usage{
soNote_flag(data = soNote_data(), field_main = "FChartNo", key = "P203031A112")
}
\arguments{
\item{data}{数据}

\item{field_main}{值}

\item{key}{主键}
}
\value{
返回值
}
\description{
自动判断相应的数据
}
\examples{
soNote_flag()
}
