% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soNote.R
\name{soNote_getMaxCalcNo}
\alias{soNote_getMaxCalcNo}
\title{获取运算单最大号}
\usage{
soNote_getMaxCalcNo(conn = tsda::conn_rds("lcrds"))
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
获取运算单最大号
}
\examples{
soNote_getMaxCalcNo()
}
