% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soNote.R
\name{soNote_uploadDB}
\alias{soNote_uploadDB}
\title{订单备注上传数据库}
\usage{
soNote_uploadDB(
  file = "data-raw/订单备注模板.xlsx",
  table_name = "t_lcrds_soNote",
  conn = tsda::conn_rds("lcrds")
)
}
\arguments{
\item{file}{文件名}

\item{table_name}{表名}

\item{conn}{连接}
}
\value{
返回值
}
\description{
订单备注上传数据库
}
\examples{
soNote_uploadDB()
}
