% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soNote.R
\name{soNote_view_query}
\alias{soNote_view_query}
\title{订单备注信息视图查询----}
\usage{
soNote_view_query(
  conn = tsda::conn_rds("lcrds"),
  FSoNo1 = "117583137",
  FSoNo2 = "117583139",
  FCalcNo = soNote_getMaxCalcNo()
)
}
\arguments{
\item{conn}{连接}

\item{FSoNo1}{开始外部订单号}

\item{FSoNo2}{结束外部订单号}

\item{FCalcNo}{运算号}
}
\value{
返回值
}
\description{
订单备注信息视图查询----
}
\examples{
soNote_view_query()
}
