% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataStat.R
\name{variable_getMode}
\alias{variable_getMode}
\title{获取变量的业务模型}
\usage{
variable_getMode(
  conn = tsda::conn_rds("lcrds"),
  FkeyNo_tag = 0,
  FLtab_tag = 0,
  FQty_tag = 0
)
}
\arguments{
\item{conn}{连接}

\item{FkeyNo_tag}{G番标记}

\item{FLtab_tag}{L番标记}

\item{FQty_tag}{Q标记}
}
\value{
返回值
}
\description{
获取变量的业务模型
}
\examples{
variable_getMode()
}
