% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataStat.R
\name{variable_search}
\alias{variable_search}
\title{针对数量进行处理}
\usage{
variable_search(key, searchTable, returnType = "int")
}
\arguments{
\item{key}{关键词}

\item{searchTable}{表}

\item{returnType}{反回值类型}
}
\value{
返回值
}
\description{
针对数量进行处理
}
\examples{
variable_search()
}
