% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataStat.R
\name{varList_sort}
\alias{varList_sort}
\title{针对变更进行排序}
\usage{
varList_sort(
  varList = "L33,L57,L65,L67,L73,L92,L99,L82",
  desc = FALSE,
  sep = ","
)
}
\arguments{
\item{varList}{变量}

\item{desc}{是否降序}

\item{sep}{分割符}
}
\value{
返回值
}
\description{
针对变更进行排序
}
\examples{
varList_sort()
}
