% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gtable.R
\name{Gtab_batchWrite_db}
\alias{Gtab_batchWrite_db}
\title{G表批量写入数据库}
\usage{
Gtab_batchWrite_db(
  conn = tsda::conn_rds("lcrds"),
  file = "./data-raw/bom_src.xlsx",
  exclude_sheetName = "YE604B797",
  show_progress = FALSE,
  max_row = 98,
  rm.na = TRUE
)
}
\arguments{
\item{conn}{连接}

\item{file}{文件}

\item{exclude_sheetName}{排除表页签}

\item{show_progress}{是否显示进度}

\item{max_row}{最大行}

\item{rm.na}{删除空行}
}
\value{
返回值
}
\description{
G表批量写入数据库
}
\examples{
Gtab_batchWrite_db()
}
